Await fail with reason (PoC)
